requirements.txt
# FastAPI and server
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.6

# HTTP client
httpx==0.28.1

# Database
psycopg[binary]==3.2.6
supabase==2.12.0

# Authentication
pyjwt==2.10.1

# Environment and utilities
python-dotenv==1.0.1

# Cortex Hybrid RAG dependencies
openai>=1.0.0
langchain-text-splitters>=0.0.1
graphiti-core>=0.3.0
qdrant-client>=1.6.0
neo4j>=5.14.0
nest-asyncio>=1.5.0
